requirements.txt
termcolor
halo
openai
python-dotenv